ai-services/requirements.txt
# AI Services Requirements
google-generativeai==0.3.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Additional AI/ML for custom models
torch>=2.2.0
transformers>=4.36.0
scikit-learn>=1.3.2
numpy>=1.24.4
pandas>=2.1.4

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2